Debug: new 'force independence' switch
This will force Substratum to work on its own, regardless of the state of
old masquerade / interfacer. It should help with our internal debugging.

It will hide itself on release builds
It will change the interfacer summary appropriately
It should **NOT* (ahem) break anything

Also some refactoring in strings.

Combined work with iskandar1023.
- used recreate
- fixed if statements